+++ structurizr-dsl/src/main/java/com/structurizr/dsl/NameValuePair.java
@@ -32,6 +32,7 @@ class NameValuePair {
 enum NameValueType {
 
     Constant,
-    Variable
+    Variable,
+    TextBlock
 
 }