+++ structurizr-import/src/main/java/com/structurizr/importer/diagrams/kroki/KrokiImporter.java
@@ -3,6 +3,7 @@ package com.structurizr.importer.diagrams.kroki;
 import com.structurizr.importer.diagrams.AbstractDiagramImporter;
 import com.structurizr.util.ImageUtils;
 import com.structurizr.util.StringUtils;
+import com.structurizr.view.ColorScheme;
 import com.structurizr.view.ImageView;
 
 import java.io.File;
@@ -17,13 +18,21 @@ public class KrokiImporter extends AbstractDiagramImporter {
     public static final String KROKI_INLINE_PROPERTY = "kroki.inline";
 
     public void importDiagram(ImageView view, String format, File file) throws Exception {
+        importDiagram(view, format, file, null);
+    }
+
+    public void importDiagram(ImageView view, String format, File file, ColorScheme colorScheme) throws Exception {
         String content = new String(Files.readAllBytes(file.toPath()), StandardCharsets.UTF_8);
         view.setTitle(file.getName());
 
-        importDiagram(view, format, content);
+        importDiagram(view, format, content, colorScheme);
     }
 
     public void importDiagram(ImageView view, String format, String content) throws Exception {
+        importDiagram(view, format, content, null);
+    }
+
+    public void importDiagram(ImageView view, String format, String content, ColorScheme colorScheme) throws Exception {
         String krokiServer = getViewOrViewSetProperty(view, KROKI_URL_PROPERTY);
         if (StringUtils.isNullOrEmpty(krokiServer)) {
             throw new IllegalArgumentException("Please define a view/viewset property named " + KROKI_URL_PROPERTY + " to specify your Kroki server");
@@ -44,12 +53,12 @@ public class KrokiImporter extends AbstractDiagramImporter {
         String inline = getViewOrViewSetProperty(view, KROKI_INLINE_PROPERTY);
         if ("true".equals(inline)) {
             if (imageFormat.equals(SVG_FORMAT)) {
-                view.setContent(ImageUtils.getSvgAsDataUri(new URL(url), true));
+                view.setContent(ImageUtils.getSvgAsDataUri(new URL(url), true), colorScheme);
             } else {
-                view.setContent(ImageUtils.getPngAsDataUri(new URL(url), true));
+                view.setContent(ImageUtils.getPngAsDataUri(new URL(url), true), colorScheme);
             }
         } else {
-            view.setContent(url);
+            view.setContent(url, colorScheme);
         }
         view.setContentType(CONTENT_TYPES_BY_FORMAT.get(imageFormat));
     }