+++ structurizr-import/src/main/java/com/structurizr/importer/diagrams/mermaid/MermaidImporter.java
@@ -3,6 +3,7 @@ package com.structurizr.importer.diagrams.mermaid;
 import com.structurizr.importer.diagrams.AbstractDiagramImporter;
 import com.structurizr.util.ImageUtils;
 import com.structurizr.util.StringUtils;
+import com.structurizr.view.ColorScheme;
 import com.structurizr.view.ImageView;
 
 import java.io.File;
@@ -18,13 +19,21 @@ public class MermaidImporter extends AbstractDiagramImporter {
     public static final String MERMAID_INLINE_PROPERTY = "mermaid.inline";
 
     public void importDiagram(ImageView view, File file) throws Exception {
+        importDiagram(view, file, null);
+    }
+
+    public void importDiagram(ImageView view, File file, ColorScheme colorScheme) throws Exception {
         String content = new String(Files.readAllBytes(file.toPath()), StandardCharsets.UTF_8);
         view.setTitle(file.getName());
 
-        importDiagram(view, content);
+        importDiagram(view, content, colorScheme);
     }
 
     public void importDiagram(ImageView view, String content) throws Exception {
+        importDiagram(view, content, null);
+    }
+
+    public void importDiagram(ImageView view, String content, ColorScheme colorScheme) throws Exception {
         String mermaidServer = getViewOrViewSetProperty(view, MERMAID_URL_PROPERTY);
         if (StringUtils.isNullOrEmpty(mermaidServer)) {
             throw new IllegalArgumentException("Please define a view/viewset property named " + MERMAID_URL_PROPERTY + " to specify your Mermaid server");
@@ -55,12 +64,12 @@ public class MermaidImporter extends AbstractDiagramImporter {
         String inline = getViewOrViewSetProperty(view, MERMAID_INLINE_PROPERTY);
         if ("true".equals(inline)) {
             if (format.equals(SVG_FORMAT)) {
-                view.setContent(ImageUtils.getSvgAsDataUri(new URL(url), true));
+                view.setContent(ImageUtils.getSvgAsDataUri(new URL(url), true), colorScheme);
             } else {
-                view.setContent(ImageUtils.getPngAsDataUri(new URL(url), true));
+                view.setContent(ImageUtils.getPngAsDataUri(new URL(url), true), colorScheme);
             }
         } else {
-            view.setContent(url);
+            view.setContent(url, colorScheme);
         }
         view.setContentType(CONTENT_TYPES_BY_FORMAT.get(format));
     }